refactoring imports and names of disclosure of electronic document pages
